[bugfix] Fix integration tests
